+++ docs/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.28.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-docs@0.28.6...@pluralsight/ps-design-system-docs@0.28.7) (2021-06-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-docs
+
+
+
+
+
+## [0.28.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-docs@0.28.5...@pluralsight/ps-design-system-docs@0.28.6) (2021-06-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-docs
+
+
+
+
+
+## [0.28.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-docs@0.28.4...@pluralsight/ps-design-system-docs@0.28.5) (2021-06-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-docs
+
+
+
+
+
+## [0.28.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-docs@0.28.3...@pluralsight/ps-design-system-docs@0.28.4) (2021-06-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-docs
+
+
+
+
+
+## [0.28.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-docs@0.28.2...@pluralsight/ps-design-system-docs@0.28.3) (2021-06-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-docs
+
+
+
+
+
 ## [0.28.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-docs@0.28.1...@pluralsight/ps-design-system-docs@0.28.2) (2021-06-08)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-docs