+++ packages/button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [22.1.9](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-button@22.1.8...@pluralsight/ps-design-system-button@22.1.9) (2021-06-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-button
+
+
+
+
+
+## [22.1.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-button@22.1.7...@pluralsight/ps-design-system-button@22.1.8) (2021-06-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-button
+
+
+
+
+
 ## [22.1.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-button@22.1.6...@pluralsight/ps-design-system-button@22.1.7) (2021-06-08)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-button